- gestion de la vitesse en fonction du terrain et du ralentissement - ajustement de l'élasticité ______ - rajout pour ce soit plus beau - rajout d'animation, etc. ------ tous les effets ont ete rajoute ------ - ajout d'un retour de force lors de la frappe contre le palet dans toutes les directions, revision de la classe Controleur qui est maintenant Observable et PaletHaptique Observer ------ - ajout reussi d'une texture ------ - tentative ajout d'autre effet sans resultat - REamélioration de l'architecture logiciel ------ - tentative ajout d'autre effet sans resultat - REamélioration de l'architecture logiciel ------ - ajout de l'effet enclos ------ - gestion du score ------ - faire l'heritage sur le palet haptique, trouver les effets pour les autres evt haptique ! attention sur l'heritage du PaletHaptique il est aussi Observer ------ - préparation a l'ajout d'effet depuis un fichier ------ - interaction avec le palet, gestion des collisions mvc appliqué - ajout des buts - ajout de quelque detail terrain ------ - base du projet sans haptique - la raquette joueur se déplace

+++ ProjetVide/But.cpp
@@ -1,6 +1,6 @@
 #include "But.h"
 
-But::But(const int &x, const int &y, const int &width) : Element(x, y, width, 0)
+But::But(const double &x, const double &y, const double &width) : Element(x, y, width, 0)
 {
 }
